Working version of plugin dependencies.
The configuration now contains one additional item for each plugin, inputFrom. It should be set to a plugin name or to 'versions' if the plugin simply wants the old and new versions. The plugin interface was updated to pass a Map instead of two strings and it's now up to the plugin itself to handle the passed map.
This will likely become problematic later on unless we introduce some type of versioning in this map.

Reflection is now removed as well and which plugin to instantiate
belongs with the configuration instead. There are no real safe guards in
place so try and play nice with the configuration.